+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Add OpenSearch URl as an optional parameter for tool calls ([#20](https://github.com/opensearch-project/opensearch-mcp-server-py/pull/20))
 - Add CI to run unit tests ([#22](https://github.com/opensearch-project/opensearch-mcp-server-py/pull/22))
 - Add support for AWS OpenSearch serverless ([#31](https://github.com/opensearch-project/opensearch-mcp-server-py/pull/31))
+- Add filtering tools based on OpenSearch version compatibility defined in TOOL_REGISTRY ([32](https://github.com/opensearch-project/opensearch-mcp-server-py/pull/32))
+
 ### Removed
 
 ### Fixed